requirements.txt
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
streamlit==1.16.0
